battle is working - enemies are spawning attack is possible, hit points are dropping. Enemy Orc class is created

+++ myRPG/Character.cpp
@@ -45,7 +45,7 @@ short int Character::getInitiative(){
 	return initiative;
 }
 
-double Character::getDamage() {
+int Character::getDamage() {
 	return damage;
 }
 
@@ -82,7 +82,7 @@ void Character::setInitiative(short int init){
 	initiative = init;
 }
 
-void Character::setDamage(double dmg) {
+void Character::setDamage(int dmg) {
 	damage = dmg;
 }
